+++ src/Yuppi.class.ts
@@ -24,7 +24,7 @@ export class Yuppi {
   public validate(schema: Schema, properties: AnyObject) {
     const yup_schema = this.convertToYup(schema);
 
-    return yup_schema.validateSync(properties, this.options.validate_options);
+    return yup_schema.validate(properties, this.options.validate_options);
   }
 
   public declare(schema: Schema, name: string) {