+++ tests/Number/Type.test.ts
@@ -24,7 +24,7 @@ const faulty_properties = [
 
 for (let i = 0; i < correct_properties.length; i++) {
   try {
-    Yupp.validate(schema, correct_properties[i]);
+    await Yupp.validate(schema, correct_properties[i]);
 
     console.log(`✅ Success ${(i + 1).toString()}/${correct_properties.length.toString()} [CORRECT_PROPERTIES]`);
   } catch {
@@ -34,7 +34,7 @@ for (let i = 0; i < correct_properties.length; i++) {
 
 for (let i = 0; i < faulty_properties.length; i++) {
   try {
-    Yupp.validate(schema, faulty_properties[i]);
+    await Yupp.validate(schema, faulty_properties[i]);
 
     throw new Error(`❌ Error ${(i + 1).toString()}/${faulty_properties.length.toString()} [FAULTY_PROPERTIES]`);
   } catch (error) {